[P<-46] FEAT: item and image

+++ src/services/item.service.js
@@ -1,6 +1,11 @@
-import { NotFoundException } from "../common/exceptions/index.js";
+import {
+  NotFoundException,
+  UnAuthorizedException,
+} from "../common/exceptions/index.js";
 import { NOT_FOUND_ITEM } from "../common/constants/index.js";
 import { itemRepository } from "../repositories/item.repository.js";
+import fs from "fs/promises";
+import path from "path";
 
 // "di"
 export class ItemService {
@@ -33,13 +38,30 @@ export class ItemService {
     return item;
   }
 
-  async createItem(data) {
-    return await this.itemRepository.create(data);
+  async createItem({ name, description, price, tags, images, userId }) {
+    const imageUrls = images.map((file) => file.path);
+
+    try {
+      const newItem = await this.itemRepository.create({
+        name,
+        description,
+        price,
+        tags,
+        images: imageUrls,
+        userId,
+      });
+      return newItem;
+    } catch (error) {
+      if (images && images.length > 0) {
+        return this.rollbackUploadFiles(images);
+      }
+    }
   }
 
   async updateItem(itemId, data) {
     try {
-      return await this.itemRepository.update(itemId, data);
+      const updatedItem = await this.itemRepository.update(itemId, data);
+      return updatedItem;
     } catch (error) {
       if (error.code === "P2025") {
         throw new NotFoundException(NOT_FOUND_ITEM);
@@ -48,16 +70,43 @@ export class ItemService {
     }
   }
 
-  async deleteItem(itemId) {
+  async deleteItem(itemId, currentUserId) {
+    const item = await this.itemRepository.findById(itemId);
+    if (!item) {
+      return new NotFoundException(NOT_FOUND_ITEM);
+    }
+
+    if (item.authorId !== currentUserId) {
+      throw new UnAuthorizedException();
+    }
+
     try {
-      const deletedItem = await this.itemRepository.delete(itemId);
-      return deletedItem;
+      await this.itemRepository.delete(itemId);
     } catch (error) {
       if (error.code === "P2025") {
         throw new NotFoundException(NOT_FOUND_ITEM);
       }
       throw error;
     }
+  }
+
+  // ---- helper ----
+  async rollbackUploadFiles(imageUrl) {
+    const deleteUrl = imageUrl.map(async (url) => {
+      const fileName = url.replace("/uploads/", "");
+      const filePath = path.join(process.cwd(), "uploads", fileName);
+
+      try {
+        await fs.unlink(filePath);
+        console.log(`롤백 - 파일을 삭제:${filePath}`);
+      } catch (unlinkError) {
+        console.error(
+          `롤백에러 - 파일 삭제를 실패함: ${filePath}`,
+          unlinkError
+        );
+      }
+    });
+    await Promise.all(deleteUrl);
   }
 }
